moved CoAPAirClientBase to base_client.py
